Move the new plugin over
